backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
openai>=1.51.0
httpx==0.27.2
pyyaml==6.0.2
